refactor: 검증 책임을 객체로 변경하기 위한 테스트 작성
