+++ src/test/java/racingcar/domain/AttemptTest.java
@@ -6,9 +6,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class AttemptTest {
 
-    @Test
-    void 같은_숫자_테스트() {
-        Attempt attempt = new Attempt(5);
-        assertTrue(attempt.isSame(5));
-    }
+	@Test
+	void 같은_숫자_테스트() {
+		Attempt attempt = new Attempt("5");
+		assertTrue(attempt.isSame(5));
+	}
 }
